+++ Libplanet.Tests/Net/Messages/NetMQMessageCodecTest.cs
@@ -9,8 +9,8 @@ using Libplanet.Crypto;
 using Libplanet.Net;
 using Libplanet.Net.Messages;
 using Libplanet.Store;
+using Libplanet.Store.Trie;
 using Libplanet.Tests.Common.Action;
-using Libplanet.Tests.Store.Trie;
 using NetMQ;
 using Xunit;